Struggling with Codes

+++ Assets/Scripts/Bullet.cs
@@ -4,7 +4,7 @@ public class Bullet : MonoBehaviour
 {
     public Vector2 direction = new Vector2(1, 0);
     public float speed = 2;
-    public float damage = 10f; // Merminin vereceði hasar
+    public int damage = 10; // Merminin vereceði hasar
 
     public Vector2 velocity;
